requirements.txt
aiohttp==3.7.4.post0
aioredis==2.0.1
async-timeout==4.0.2
attrs==21.4.0
chardet==4.0.0
cleo==1.0.0a4
click==8.1.3
colorama==0.4.5
crashtest==0.3.1
Deprecated==1.2.13
discord.py==1.7.3
hiredis==2.0.0
idna==3.3
multidict==6.0.2
packaging==21.3
pptree==3.1
pydantic==1.9.1
pylev==1.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-ulid==1.1.0
pytz==2021.3
redis==4.3.3
redis-om==0.0.27
six==1.16.0
types-redis==4.3.2
types-six==1.16.16
typing_extensions==4.2.0
wrapt==1.14.1
yarl==1.7.2
